requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
langchain>=0.0.339
langchain-community>=0.0.16
langchain-core>=0.1.16
langchain-pinecone>=0.0.1
groq==0.4.2
chromadb==0.4.18
pymupdf==1.22.5
spacy==3.7.2
streamlit==1.28.0
pypdf==3.17.0
python-multipart==0.0.6
pandas==2.1.1
markdown==3.5
numpy==1.26.1
sentence-transformers==2.2.2
transformers>=4.36.0
torch>=2.0.0
pytest==7.4.3
httpx==0.25.1
regex==2023.10.3
llama-index==0.8.54
pinecone-client==2.2.4 